+++ Changes.md
@@ -7,6 +7,9 @@ _A Detailed Changelog from the last Release is available [here](https://gitlab.c
 ### Changes
 
 *   Added `/cp export assets` command (Has two optional arguments for usage specification)
+*   (Breaking) Changed `Show Game State` to `Detect World Data` to be more representative
+*   (Breaking) Changed `Show Biome` and `Show Dimension` to `Detect Biome Data` and `Detect Dimension Data` to be more representative
+*   List Data will no longer update before the config is saved (Related to v1.7.2's change)
 
 ### Fixes